GH-18 Refactor pagenum assignment and history push
Place functions for pagenum assignment and history push in utils
other components can use these methods to manipulate history and
retrieve search params
